BL33: v2019: clean up warning [1/1]
PD#SWPL-43527

Problem:
need to enable Werror

Solution:
initial cleanup for compiling warning

Verify:
./check_compile all chip

+++ arch/arm/include/asm/io.h
@@ -41,12 +41,12 @@ static inline void sync(void)
  */
 #define __arch_getb(a)			(*(volatile unsigned char *)(a))
 #define __arch_getw(a)			(*(volatile unsigned short *)(a))
-#define __arch_getl(a)			(*(volatile unsigned int *)(a))
+#define __arch_getl(a)			(*(volatile unsigned int *)(uintptr_t)(a))
 #define __arch_getq(a)			(*(volatile unsigned long long *)(a))
 
 #define __arch_putb(v,a)		(*(volatile unsigned char *)(a) = (v))
 #define __arch_putw(v,a)		(*(volatile unsigned short *)(a) = (v))
-#define __arch_putl(v,a)		(*(volatile unsigned int *)(a) = (v))
+#define __arch_putl(v,a)		(*(volatile unsigned int *)(uintptr_t)(a) = (v))
 #define __arch_putq(v,a)		(*(volatile unsigned long long *)(a) = (v))
 
 static inline void __raw_writesb(unsigned long addr, const void *data,
